# Conflicts:
#	texmap/scanline.cpp

+++ 2d/scanline2d.cpp
@@ -31,15 +31,18 @@ void gr_linear_stosd(uint8_t* dest, uint8_t color, int count)
 {
 	int i, x;
 
-	if (count > 3) {
-		while ((int)(dest) & 0x3) { *dest++ = color; count--; };
-		if (count >= 4) {
+	if (count > 3) 
+	{
+		while ((uintptr_t)(dest) & 0x3) { *dest++ = color; count--; };
+		if (count >= 4) 
+		{
 			x = (color << 24) | (color << 16) | (color << 8) | color;
 			while (count > 4) { *(int*)dest = x; dest += 4; count -= 4; };
 		}
 		while (count > 0) { *dest++ = color; count--; };
 	}
-	else {
+	else 
+	{
 		for (i = 0; i < count; i++)
 			* dest++ = color;
 	}
