+++ 3d/points.cpp
@@ -57,10 +57,10 @@ uint8_t g3_rotate_point(g3s_point* dest, vms_vector* src)
 //returns true if div is ok, else false
 int checkmuldiv(fix* r, fix a, fix b, fix c)
 {
-	long long q, qt;
+	int64_t q, qt;
 	q = 0; qt = 0;
 
-	long high; uint32_t low;
+	int32_t high; uint32_t low;
 
 	//q.low = q.high = 0;
 	fixmulaccum(&q, a, b);
@@ -186,7 +186,7 @@ uint8_t g3_add_delta_vec(g3s_point* dest, g3s_point* src, vms_vector* deltav)
 //calculate the depth of a point - returns the z coord of the rotated point
 fix g3_calc_point_depth(vms_vector* pnt)
 {
-	long long q;
+	int64_t q;
 
 	q = 0;
 	fixmulaccum(&q, (pnt->x - View_position.x), View_matrix.fvec.x);
